Refactor ConfigHandle
