+++ packages/realm_dart/lib/src/native/realm_core.dart
@@ -41,6 +41,7 @@ import 'realm_bindings.dart';
 import 'realm_library.dart';
 
 // TODO: Use regular
+part 'config_handle.dart';
 part 'convert.dart';
 part 'decimal128.dart';
 part 'error_handling.dart';
@@ -74,6 +75,97 @@ final _pluginLib = () {
 _RealmCore realmCore = _RealmCore();
 
 const encryptionKeySize = 64;
+
+void _guardSynchronousCallback(FutureOr<void> Function() callback, Pointer<Void> unlockCallbackFunc) async {
+  Pointer<Void> user_error = nullptr;
+  try {
+    await callback();
+  } catch (error) {
+    user_error = error.toPersistentHandle();
+  } finally {
+    realmLib.realm_dart_invoke_unlock_callback(user_error, unlockCallbackFunc);
+  }
+}
+
+bool should_compact_callback(Pointer<Void> userdata, int totalSize, int usedSize) {
+  final config = userdata.toObject();
+
+  if (config is LocalConfiguration) {
+    return config.shouldCompactCallback!(totalSize, usedSize);
+  } else if (config is FlexibleSyncConfiguration) {
+    return config.shouldCompactCallback!(totalSize, usedSize);
+  }
+
+  return false;
+}
+
+bool migration_callback(Pointer<Void> userdata, Pointer<shared_realm> oldRealmHandle, Pointer<shared_realm> newRealmHandle, Pointer<realm_schema> schema) {
+  final oldHandle = RealmHandle._unowned(oldRealmHandle);
+  final newHandle = RealmHandle._unowned(newRealmHandle);
+  try {
+    final LocalConfiguration config = userdata.toObject();
+
+    final oldSchemaVersion = realmLib.realm_get_schema_version(oldRealmHandle);
+    final oldConfig = Configuration.local([], path: config.path, isReadOnly: true, schemaVersion: oldSchemaVersion);
+    final oldRealm = RealmInternal.getUnowned(oldConfig, oldHandle, isInMigration: true);
+
+    final newRealm = RealmInternal.getUnowned(config, newHandle, isInMigration: true);
+
+    final migration = MigrationInternal.create(RealmInternal.getMigrationRealm(oldRealm), newRealm, SchemaHandle.unowned(schema));
+    config.migrationCallback!(migration, oldSchemaVersion);
+    return true;
+  } catch (ex) {
+    realmLib.realm_register_user_code_callback_error(ex.toPersistentHandle());
+  } finally {
+    oldHandle.release();
+    newHandle.release();
+  }
+
+  return false;
+}
+
+void _syncErrorHandlerCallback(Object userdata, Pointer<realm_sync_session> session, realm_sync_error error) {
+  final syncConfig = userdata as FlexibleSyncConfiguration;
+  // TODO: Take the app from the session instead of from syncConfig after fixing issue https://github.com/realm/realm-dart/issues/633
+  final syncError = SyncErrorInternal.createSyncError(error.toDart(), app: syncConfig.user.app);
+
+  if (syncError is ClientResetError) {
+    syncConfig.clientResetHandler.onManualReset?.call(syncError);
+    return;
+  }
+
+  syncConfig.syncErrorHandler(syncError);
+}
+
+void _syncBeforeResetCallback(Object userdata, Pointer<shared_realm> realmHandle, Pointer<Void> unlockCallbackFunc) {
+  _guardSynchronousCallback(() async {
+    final syncConfig = userdata as FlexibleSyncConfiguration;
+    var beforeResetCallback = syncConfig.clientResetHandler.onBeforeReset!;
+
+    final realm = RealmInternal.getUnowned(syncConfig, RealmHandle._unowned(realmHandle));
+    try {
+      await beforeResetCallback(realm);
+    } finally {
+      realm.handle.release();
+    }
+  }, unlockCallbackFunc);
+}
+
+bool initial_data_callback(Pointer<Void> userdata, Pointer<shared_realm> realmPtr) {
+  final realmHandle = RealmHandle._unowned(realmPtr);
+  try {
+    final LocalConfiguration config = userdata.toObject();
+    final realm = RealmInternal.getUnowned(config, realmHandle);
+    config.initialDataCallback!(realm);
+    return true;
+  } catch (ex) {
+    realmLib.realm_register_user_code_callback_error(ex.toPersistentHandle());
+  } finally {
+    realmHandle.release();
+  }
+
+  return false;
+}
 
 // All access to Realm Core functionality goes through this class
 class _RealmCore {
@@ -175,132 +267,6 @@ class _RealmCore {
     });
   }
 
-  static ConfigHandle _createConfig(Configuration config) {
-    return using((Arena arena) {
-      final configPtr = realmLib.realm_config_new();
-      final configHandle = ConfigHandle._(configPtr);
-
-      if (config.schemaObjects.isNotEmpty) {
-        final schemaHandle = _createSchema(config.schemaObjects);
-        realmLib.realm_config_set_schema(configHandle.pointer, schemaHandle.pointer);
-      }
-
-      realmLib.realm_config_set_path(configHandle.pointer, config.path.toCharPtr(arena));
-      realmLib.realm_config_set_scheduler(configHandle.pointer, scheduler.handle.pointer);
-
-      if (config.fifoFilesFallbackPath != null) {
-        realmLib.realm_config_set_fifo_path(configHandle.pointer, config.fifoFilesFallbackPath!.toCharPtr(arena));
-      }
-
-      // Setting schema version only makes sense for local realms, but core insists it is always set,
-      // hence we set it to 0 in those cases.
-
-      final schemaVersion = switch (config) {
-        (LocalConfiguration lc) => lc.schemaVersion,
-        (FlexibleSyncConfiguration fsc) => fsc.schemaVersion,
-        _ => 0,
-      };
-      realmLib.realm_config_set_schema_version(configHandle.pointer, schemaVersion);
-      if (config.maxNumberOfActiveVersions != null) {
-        realmLib.realm_config_set_max_number_of_active_versions(configHandle.pointer, config.maxNumberOfActiveVersions!);
-      }
-      if (config is LocalConfiguration) {
-        //realmLib.realm_config_set_schema_mode(configHandle.pointer, realm_schema_mode.RLM_SCHEMA_MODE_ADDITIVE_DISCOVERED);
-        if (config.initialDataCallback != null) {
-          realmLib.realm_config_set_data_initialization_function(
-            configHandle.pointer,
-            Pointer.fromFunction(initial_data_callback, false),
-            config.toPersistentHandle(),
-            realmLib.addresses.realm_dart_delete_persistent_handle,
-          );
-        }
-        if (config.isReadOnly) {
-          realmLib.realm_config_set_schema_mode(configHandle.pointer, realm_schema_mode.RLM_SCHEMA_MODE_IMMUTABLE);
-        } else if (config.shouldDeleteIfMigrationNeeded) {
-          realmLib.realm_config_set_schema_mode(configHandle.pointer, realm_schema_mode.RLM_SCHEMA_MODE_SOFT_RESET_FILE);
-        }
-        if (config.disableFormatUpgrade) {
-          realmLib.realm_config_set_disable_format_upgrade(configHandle.pointer, config.disableFormatUpgrade);
-        }
-        if (config.shouldCompactCallback != null) {
-          realmLib.realm_config_set_should_compact_on_launch_function(
-            configHandle.pointer,
-            Pointer.fromFunction(should_compact_callback, false),
-            config.toPersistentHandle(),
-            realmLib.addresses.realm_dart_delete_persistent_handle,
-          );
-        }
-        if (config.migrationCallback != null) {
-          realmLib.realm_config_set_migration_function(
-            configHandle.pointer,
-            Pointer.fromFunction(migration_callback, false),
-            config.toPersistentHandle(),
-            realmLib.addresses.realm_dart_delete_persistent_handle,
-          );
-        }
-      } else if (config is InMemoryConfiguration) {
-        realmLib.realm_config_set_in_memory(configHandle.pointer, true);
-      } else if (config is FlexibleSyncConfiguration) {
-        realmLib.realm_config_set_schema_mode(configHandle.pointer, realm_schema_mode.RLM_SCHEMA_MODE_ADDITIVE_DISCOVERED);
-        final syncConfigPtr = invokeGetPointer(() => realmLib.realm_flx_sync_config_new(config.user.handle.pointer));
-        try {
-          realmLib.realm_sync_config_set_session_stop_policy(syncConfigPtr, config.sessionStopPolicy.index);
-          realmLib.realm_sync_config_set_resync_mode(syncConfigPtr, config.clientResetHandler.clientResyncMode.index);
-          final errorHandlerCallback =
-              Pointer.fromFunction<Void Function(Handle, Pointer<realm_sync_session_t>, realm_sync_error_t)>(_syncErrorHandlerCallback);
-          final errorHandlerUserdata = realmLib.realm_dart_userdata_async_new(config, errorHandlerCallback.cast(), scheduler.handle.pointer);
-          realmLib.realm_sync_config_set_error_handler(syncConfigPtr, realmLib.addresses.realm_dart_sync_error_handler_callback, errorHandlerUserdata.cast(),
-              realmLib.addresses.realm_dart_userdata_async_free);
-
-          if (config.clientResetHandler.onBeforeReset != null) {
-            final syncBeforeResetCallback = Pointer.fromFunction<Void Function(Handle, Pointer<shared_realm>, Pointer<Void>)>(_syncBeforeResetCallback);
-            final beforeResetUserdata = realmLib.realm_dart_userdata_async_new(config, syncBeforeResetCallback.cast(), scheduler.handle.pointer);
-
-            realmLib.realm_sync_config_set_before_client_reset_handler(syncConfigPtr, realmLib.addresses.realm_dart_sync_before_reset_handler_callback,
-                beforeResetUserdata.cast(), realmLib.addresses.realm_dart_userdata_async_free);
-          }
-
-          if (config.clientResetHandler.onAfterRecovery != null || config.clientResetHandler.onAfterDiscard != null) {
-            final syncAfterResetCallback =
-                Pointer.fromFunction<Void Function(Handle, Pointer<shared_realm>, Pointer<realm_thread_safe_reference>, Bool, Pointer<Void>)>(
-                    _syncAfterResetCallback);
-            final afterResetUserdata = realmLib.realm_dart_userdata_async_new(config, syncAfterResetCallback.cast(), scheduler.handle.pointer);
-
-            realmLib.realm_sync_config_set_after_client_reset_handler(syncConfigPtr, realmLib.addresses.realm_dart_sync_after_reset_handler_callback,
-                afterResetUserdata.cast(), realmLib.addresses.realm_dart_userdata_async_free);
-          }
-
-          if (config.shouldCompactCallback != null) {
-            realmLib.realm_config_set_should_compact_on_launch_function(
-              configHandle.pointer,
-              Pointer.fromFunction(should_compact_callback, false),
-              config.toPersistentHandle(),
-              realmLib.addresses.realm_dart_delete_persistent_handle,
-            );
-          }
-
-          realmLib.realm_config_set_sync_config(configPtr, syncConfigPtr);
-        } finally {
-          realmLib.realm_release(syncConfigPtr.cast());
-        }
-      } else if (config is DisconnectedSyncConfiguration) {
-        realmLib.realm_config_set_schema_mode(configHandle.pointer, realm_schema_mode.RLM_SCHEMA_MODE_ADDITIVE_EXPLICIT);
-        realmLib.realm_config_set_force_sync_history(configPtr, true);
-      }
-
-      if (config.encryptionKey != null) {
-        realmLib.realm_config_set_encryption_key(configPtr, config.encryptionKey!.toUint8Ptr(arena), encryptionKeySize);
-      }
-
-      // For sync and for dynamic Realms, we need to have a complete view of the schema in Core.
-      if (config.schemaObjects.isEmpty || config is FlexibleSyncConfiguration) {
-        realmLib.realm_config_set_schema_subset_mode(configHandle.pointer, realm_schema_subset_mode.RLM_SCHEMA_SUBSET_MODE_COMPLETE);
-      }
-
-      return configHandle;
-    });
-  }
-
   String getPathForUser(User user) {
     final syncConfigPtr = invokeGetPointer(() => realmLib.realm_flx_sync_config_new(user.handle.pointer));
     try {
@@ -313,121 +279,6 @@ class _RealmCore {
 
   SubscriptionSetHandle getSubscriptions(Realm realm) {
     return SubscriptionSetHandle._(invokeGetPointer(() => realmLib.realm_sync_get_active_subscription_set(realm.handle.pointer)), realm.handle);
-  }
-
-  static bool initial_data_callback(Pointer<Void> userdata, Pointer<shared_realm> realmPtr) {
-    final realmHandle = RealmHandle._unowned(realmPtr);
-    try {
-      final LocalConfiguration config = userdata.toObject();
-      final realm = RealmInternal.getUnowned(config, realmHandle);
-      config.initialDataCallback!(realm);
-      return true;
-    } catch (ex) {
-      realmLib.realm_register_user_code_callback_error(ex.toPersistentHandle());
-    } finally {
-      realmHandle.release();
-    }
-
-    return false;
-  }
-
-  static bool should_compact_callback(Pointer<Void> userdata, int totalSize, int usedSize) {
-    final config = userdata.toObject();
-
-    if (config is LocalConfiguration) {
-      return config.shouldCompactCallback!(totalSize, usedSize);
-    } else if (config is FlexibleSyncConfiguration) {
-      return config.shouldCompactCallback!(totalSize, usedSize);
-    }
-
-    return false;
-  }
-
-  static bool migration_callback(
-      Pointer<Void> userdata, Pointer<shared_realm> oldRealmHandle, Pointer<shared_realm> newRealmHandle, Pointer<realm_schema> schema) {
-    final oldHandle = RealmHandle._unowned(oldRealmHandle);
-    final newHandle = RealmHandle._unowned(newRealmHandle);
-    try {
-      final LocalConfiguration config = userdata.toObject();
-
-      final oldSchemaVersion = realmLib.realm_get_schema_version(oldRealmHandle);
-      final oldConfig = Configuration.local([], path: config.path, isReadOnly: true, schemaVersion: oldSchemaVersion);
-      final oldRealm = RealmInternal.getUnowned(oldConfig, oldHandle, isInMigration: true);
-
-      final newRealm = RealmInternal.getUnowned(config, newHandle, isInMigration: true);
-
-      final migration = MigrationInternal.create(RealmInternal.getMigrationRealm(oldRealm), newRealm, SchemaHandle.unowned(schema));
-      config.migrationCallback!(migration, oldSchemaVersion);
-      return true;
-    } catch (ex) {
-      realmLib.realm_register_user_code_callback_error(ex.toPersistentHandle());
-    } finally {
-      oldHandle.release();
-      newHandle.release();
-    }
-
-    return false;
-  }
-
-  static void _syncErrorHandlerCallback(Object userdata, Pointer<realm_sync_session> session, realm_sync_error error) {
-    final syncConfig = userdata as FlexibleSyncConfiguration;
-    // TODO: Take the app from the session instead of from syncConfig after fixing issue https://github.com/realm/realm-dart/issues/633
-    final syncError = SyncErrorInternal.createSyncError(error.toDart(), app: syncConfig.user.app);
-
-    if (syncError is ClientResetError) {
-      syncConfig.clientResetHandler.onManualReset?.call(syncError);
-      return;
-    }
-
-    syncConfig.syncErrorHandler(syncError);
-  }
-
-  static void _guardSynchronousCallback(FutureOr<void> Function() callback, Pointer<Void> unlockCallbackFunc) async {
-    Pointer<Void> user_error = nullptr;
-    try {
-      await callback();
-    } catch (error) {
-      user_error = error.toPersistentHandle();
-    } finally {
-      realmLib.realm_dart_invoke_unlock_callback(user_error, unlockCallbackFunc);
-    }
-  }
-
-  static void _syncBeforeResetCallback(Object userdata, Pointer<shared_realm> realmHandle, Pointer<Void> unlockCallbackFunc) {
-    _guardSynchronousCallback(() async {
-      final syncConfig = userdata as FlexibleSyncConfiguration;
-      var beforeResetCallback = syncConfig.clientResetHandler.onBeforeReset!;
-
-      final realm = RealmInternal.getUnowned(syncConfig, RealmHandle._unowned(realmHandle));
-      try {
-        await beforeResetCallback(realm);
-      } finally {
-        realm.handle.release();
-      }
-    }, unlockCallbackFunc);
-  }
-
-  static void _syncAfterResetCallback(Object userdata, Pointer<shared_realm> beforeHandle, Pointer<realm_thread_safe_reference> afterReference, bool didRecover,
-      Pointer<Void> unlockCallbackFunc) {
-    _guardSynchronousCallback(() async {
-      final syncConfig = userdata as FlexibleSyncConfiguration;
-      final afterResetCallback = didRecover ? syncConfig.clientResetHandler.onAfterRecovery : syncConfig.clientResetHandler.onAfterDiscard;
-
-      if (afterResetCallback == null) {
-        return;
-      }
-
-      final beforeRealm = RealmInternal.getUnowned(syncConfig, RealmHandle._unowned(beforeHandle));
-      final realmPtr = invokeGetPointer(() => realmLib.realm_from_thread_safe_reference(afterReference, scheduler.handle.pointer));
-      final afterRealm = RealmInternal.getUnowned(syncConfig, RealmHandle._unowned(realmPtr));
-
-      try {
-        return await afterResetCallback(beforeRealm, afterRealm);
-      } finally {
-        beforeRealm.handle.release();
-        afterRealm.handle.release();
-      }
-    }, unlockCallbackFunc);
   }
 
   void raiseError(Session session, int errorCode, bool isFatal) {
@@ -452,13 +303,13 @@ class _RealmCore {
   }
 
   RealmHandle openRealm(Configuration config) {
-    final configHandle = _createConfig(config);
+    final configHandle = ConfigHandle(config);
     final realmPtr = invokeGetPointer(() => realmLib.realm_open(configHandle.pointer), "Error opening realm at path ${config.path}");
     return RealmHandle._(realmPtr);
   }
 
   RealmAsyncOpenTaskHandle createRealmAsyncOpenTask(FlexibleSyncConfiguration config) {
-    final configHandle = _createConfig(config);
+    final configHandle = ConfigHandle(config);
     final asyncOpenTaskPtr = invokeGetPointer(() => realmLib.realm_open_synchronized(configHandle.pointer), "Error opening realm at path ${config.path}");
     return RealmAsyncOpenTaskHandle._(asyncOpenTaskPtr);
   }
@@ -2804,10 +2655,7 @@ class _RealmCore {
     });
   }
 
-  void writeCopy(Realm realm, Configuration config) {
-    final configHandle = _createConfig(config);
-    invokeGetBool(() => realmLib.realm_convert_with_config(realm.handle.pointer, configHandle.pointer, false));
-  }
+  void writeCopy(Realm realm, Configuration config) => realm.handle.writeCopy(config);
 
   void _createCollection(Realm realm, RealmValue value, Pointer<realm_list> Function() createList, Pointer<realm_dictionary> Function() createMap) {
     CollectionHandleBase? collectionHandle;
@@ -2867,10 +2715,6 @@ class SchemaHandle extends HandleBase<realm_schema> {
   SchemaHandle._(Pointer<realm_schema> pointer) : super(pointer, 24);
 
   SchemaHandle.unowned(super.pointer) : super.unowned();
-}
-
-class ConfigHandle extends HandleBase<realm_config> {
-  ConfigHandle._(Pointer<realm_config> pointer) : super(pointer, 512);
 }
 
 class SchedulerHandle extends HandleBase<realm_scheduler> {
