+++ packages/realm_dart/test/configuration_test.dart
@@ -3,9 +3,10 @@
 
 import 'dart:io';
 import 'dart:math';
-import 'package:test/test.dart' hide test, throws;
+
 import 'package:path/path.dart' as path;
 import 'package:realm_dart/realm.dart';
+
 import 'test.dart';
 
 void main() {
