+++ packages/realm_dart/test/realm_test.dart
@@ -9,7 +9,6 @@ import 'dart:isolate';
 import 'package:path/path.dart' as p;
 import 'package:realm_dart/realm.dart';
 import 'package:realm_dart/src/native/realm_core.dart';
-import 'package:test/test.dart' hide test, throws;
 import 'package:timezone/data/latest.dart' as tz;
 import 'package:timezone/timezone.dart' as tz;
